pyproject.toml
[project]
name = "hestia"
version = "0.2.0"
description = "Local-first personal assistant framework for consumer GPU LLMs"
readme = "README.md"
authors = [
    { name = "Dylan O'Keefe", email = "dylanokeefedev@gmail.com" }
]
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "click>=8.1.8",
    "croniter>=2.0",
    "httpx>=0.28.1",
    "matrix-nio>=0.25.0",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-telegram-bot>=21.0",
    "sqlalchemy[asyncio]>=2.0.49",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.31.0"]

[project.scripts]
hestia = "hestia.cli:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "matrix_e2e: Real Matrix homeserver end-to-end tests (requires env vars)",
]
